Split the visualization registry mechanism out of the barrel
Move the registry Map, registration functions and accessors
(registerVisualization, registerSettingWidgets, getSettingWidgetComponent,
getVisualization, getMaxDimensionsSupported, getMaxMetricsSupported, etc.)
from metabase/visualizations/index.ts into lib/registry.ts. The barrel
re-exports everything for back-compat; registration files stay app-side.
